requirements.txt
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
openai-whisper==20230918
pykakasi==2.2.1
PySimpleGUI==4.60.5
speechbrain==0.5.15
soundfile==0.12.1
pydub==0.25.1